refactor integration tests deploy remove unnecessary code
